git_ui: Decouple git_ui from the recent_projects crate
