+++ crates/editor/src/editor.rs
@@ -27827,6 +27827,12 @@ impl ui_input::ErasedEditor for ErasedEditorImpl {
             (callback)(event, window, cx);
         })
     }
+
+    fn set_masked(&self, masked: bool, _window: &mut Window, cx: &mut App) {
+        self.0.update(cx, |editor, cx| {
+            editor.set_masked(masked, cx);
+        });
+    }
 }
 impl<T> Default for InvalidationStack<T> {
     fn default() -> Self {
