+++ crates/recent_projects/Cargo.toml
@@ -39,6 +39,7 @@ picker.workspace = true
 project.workspace = true
 release_channel.workspace = true
 remote.workspace = true
+remote_connection.workspace = true
 semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -48,6 +49,7 @@ task.workspace = true
 telemetry.workspace = true
 theme.workspace = true
 ui.workspace = true
+ui_input.workspace = true
 util.workspace = true
 workspace.workspace = true
 worktree.workspace = true
