+++ crates/recent_projects/src/recent_projects.rs
@@ -11,7 +11,8 @@ mod wsl_picker;
 
 use dev_container::start_dev_container;
 use remote::RemoteConnectionOptions;
-pub use remote_connections::{RemoteConnectionModal, connect, open_remote_project};
+pub use remote_connection::{RemoteConnectionModal, connect};
+pub use remote_connections::open_remote_project;
 
 use disconnected_overlay::DisconnectedOverlay;
 use fuzzy::{StringMatch, StringMatchCandidate};
