+++ crates/recent_projects/src/remote_connections.rs
@@ -5,33 +5,26 @@ use std::{
 
 use anyhow::{Context as _, Result};
 use askpass::EncryptedPassword;
-use auto_update::AutoUpdater;
 use editor::Editor;
 use extension_host::ExtensionStore;
 use futures::{FutureExt as _, channel::oneshot, select};
-use gpui::{
-    AnyWindowHandle, App, AsyncApp, DismissEvent, Entity, EventEmitter, Focusable, FontFeatures,
-    ParentElement as _, PromptLevel, Render, SharedString, Task, TextStyleRefinement, WeakEntity,
-};
+use gpui::{AppContext, AsyncApp, PromptLevel};
 
-use language::{CursorShape, Point};
-use markdown::{Markdown, MarkdownElement, MarkdownStyle};
+use language::Point;
 use project::trusted_worktrees;
-use release_channel::ReleaseChannel;
 use remote::{
-    ConnectionIdentifier, DockerConnectionOptions, Interactive, RemoteClient, RemoteConnection,
-    RemoteConnectionOptions, RemotePlatform, SshConnectionOptions,
+    DockerConnectionOptions, Interactive, RemoteConnection, RemoteConnectionOptions,
+    SshConnectionOptions,
 };
-use semver::Version;
 pub use settings::SshConnection;
 use settings::{DevContainerConnection, ExtendingVec, RegisterSetting, Settings, WslConnection};
-use theme::ThemeSettings;
-use ui::{
-    ActiveTheme, Color, CommonAnimationExt, Context, InteractiveElement, IntoElement, KeyBinding,
-    LabelCommon, ListItem, Styled, Window, prelude::*,
-};
 use util::paths::PathWithPosition;
-use workspace::{AppState, ModalView, Workspace};
+use workspace::{AppState, Workspace};
+
+pub use remote_connection::{
+    RemoteClientDelegate, RemoteConnectionModal, RemoteConnectionPrompt, SshConnectionHeader,
+    connect,
+};
 
 #[derive(RegisterSetting)]
 pub struct RemoteSettings {
@@ -129,505 +122,6 @@ impl Settings for RemoteSettings {
     }
 }
 
-pub struct RemoteConnectionPrompt {
-    connection_string: SharedString,
-    nickname: Option<SharedString>,
-    is_wsl: bool,
-    is_devcontainer: bool,
-    status_message: Option<SharedString>,
-    prompt: Option<(Entity<Markdown>, oneshot::Sender<EncryptedPassword>)>,
-    cancellation: Option<oneshot::Sender<()>>,
-    editor: Entity<Editor>,
-}
-
-impl Drop for RemoteConnectionPrompt {
-    fn drop(&mut self) {
-        if let Some(cancel) = self.cancellation.take() {
-            log::debug!("cancelling remote connection");
-            cancel.send(()).ok();
-        }
-    }
-}
-
-pub struct RemoteConnectionModal {
-    pub prompt: Entity<RemoteConnectionPrompt>,
-    paths: Vec<PathBuf>,
-    finished: bool,
-}
-
-impl RemoteConnectionPrompt {
-    pub(crate) fn new(
-        connection_string: String,
-        nickname: Option<String>,
-        is_wsl: bool,
-        is_devcontainer: bool,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Self {
-        Self {
-            connection_string: connection_string.into(),
-            nickname: nickname.map(|nickname| nickname.into()),
-            is_wsl,
-            is_devcontainer,
-            editor: cx.new(|cx| Editor::single_line(window, cx)),
-            status_message: None,
-            cancellation: None,
-            prompt: None,
-        }
-    }
-
-    pub fn set_cancellation_tx(&mut self, tx: oneshot::Sender<()>) {
-        self.cancellation = Some(tx);
-    }
-
-    fn set_prompt(
-        &mut self,
-        prompt: String,
-        tx: oneshot::Sender<EncryptedPassword>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let theme = ThemeSettings::get_global(cx);
-
-        let refinement = TextStyleRefinement {
-            font_family: Some(theme.buffer_font.family.clone()),
-            font_features: Some(FontFeatures::disable_ligatures()),
-            font_size: Some(theme.buffer_font_size(cx).into()),
-            color: Some(cx.theme().colors().editor_foreground),
-            background_color: Some(gpui::transparent_black()),
-            ..Default::default()
-        };
-
-        self.editor.update(cx, |editor, cx| {
-            if prompt.contains("yes/no") {
-                editor.set_masked(false, cx);
-            } else {
-                editor.set_masked(true, cx);
-            }
-            editor.set_text_style_refinement(refinement);
-            editor.set_cursor_shape(CursorShape::Block, cx);
-        });
-
-        let markdown = cx.new(|cx| Markdown::new_text(prompt.into(), cx));
-        self.prompt = Some((markdown, tx));
-        self.status_message.take();
-        window.focus(&self.editor.focus_handle(cx), cx);
-        cx.notify();
-    }
-
-    pub fn set_status(&mut self, status: Option<String>, cx: &mut Context<Self>) {
-        self.status_message = status.map(|s| s.into());
-        cx.notify();
-    }
-
-    pub fn confirm(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        if let Some((_, tx)) = self.prompt.take() {
-            self.status_message = Some("Connecting".into());
-
-            self.editor.update(cx, |editor, cx| {
-                let pw = editor.text(cx);
-                if let Ok(secure) = EncryptedPassword::try_from(pw.as_ref()) {
-                    tx.send(secure).ok();
-                }
-                editor.clear(window, cx);
-            });
-        }
-    }
-}
-
-impl Render for RemoteConnectionPrompt {
-    fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let theme = ThemeSettings::get_global(cx);
-
-        let mut text_style = window.text_style();
-        let refinement = TextStyleRefinement {
-            font_family: Some(theme.buffer_font.family.clone()),
-            font_features: Some(FontFeatures::disable_ligatures()),
-            font_size: Some(theme.buffer_font_size(cx).into()),
-            color: Some(cx.theme().colors().editor_foreground),
-            background_color: Some(gpui::transparent_black()),
-            ..Default::default()
-        };
-
-        text_style.refine(&refinement);
-        let markdown_style = MarkdownStyle {
-            base_text_style: text_style,
-            selection_background_color: cx.theme().colors().element_selection_background,
-            ..Default::default()
-        };
-
-        v_flex()
-            .key_context("PasswordPrompt")
-            .p_2()
-            .size_full()
-            .text_buffer(cx)
-            .when_some(self.status_message.clone(), |el, status_message| {
-                el.child(
-                    h_flex()
-                        .gap_2()
-                        .child(
-                            Icon::new(IconName::ArrowCircle)
-                                .color(Color::Muted)
-                                .with_rotate_animation(2),
-                        )
-                        .child(
-                            div()
-                                .text_ellipsis()
-                                .overflow_x_hidden()
-                                .child(format!("{}…", status_message)),
-                        ),
-                )
-            })
-            .when_some(self.prompt.as_ref(), |el, prompt| {
-                el.child(
-                    div()
-                        .size_full()
-                        .overflow_hidden()
-                        .child(MarkdownElement::new(prompt.0.clone(), markdown_style))
-                        .child(self.editor.clone()),
-                )
-                .when(window.capslock().on, |el| {
-                    el.child(Label::new("⚠️ ⇪ is on"))
-                })
-            })
-    }
-}
-
-impl RemoteConnectionModal {
-    pub fn new(
-        connection_options: &RemoteConnectionOptions,
-        paths: Vec<PathBuf>,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> Self {
-        let (connection_string, nickname, is_wsl, is_devcontainer) = match connection_options {
-            RemoteConnectionOptions::Ssh(options) => (
-                options.connection_string(),
-                options.nickname.clone(),
-                false,
-                false,
-            ),
-            RemoteConnectionOptions::Wsl(options) => {
-                (options.distro_name.clone(), None, true, false)
-            }
-            RemoteConnectionOptions::Docker(options) => (options.name.clone(), None, false, true),
-            #[cfg(any(test, feature = "test-support"))]
-            RemoteConnectionOptions::Mock(options) => {
-                (format!("mock-{}", options.id), None, false, false)
-            }
-        };
-        Self {
-            prompt: cx.new(|cx| {
-                RemoteConnectionPrompt::new(
-                    connection_string,
-                    nickname,
-                    is_wsl,
-                    is_devcontainer,
-                    window,
-                    cx,
-                )
-            }),
-            finished: false,
-            paths,
-        }
-    }
-
-    fn confirm(&mut self, _: &menu::Confirm, window: &mut Window, cx: &mut Context<Self>) {
-        self.prompt
-            .update(cx, |prompt, cx| prompt.confirm(window, cx))
-    }
-
-    pub fn finished(&mut self, cx: &mut Context<Self>) {
-        self.finished = true;
-        cx.emit(DismissEvent);
-    }
-
-    fn dismiss(&mut self, _: &menu::Cancel, _: &mut Window, cx: &mut Context<Self>) {
-        if let Some(tx) = self
-            .prompt
-            .update(cx, |prompt, _cx| prompt.cancellation.take())
-        {
-            log::debug!("cancelling remote connection");
-            tx.send(()).ok();
-        }
-        self.finished(cx);
-    }
-}
-
-pub(crate) struct SshConnectionHeader {
-    pub(crate) connection_string: SharedString,
-    pub(crate) paths: Vec<PathBuf>,
-    pub(crate) nickname: Option<SharedString>,
-    pub(crate) is_wsl: bool,
-    pub(crate) is_devcontainer: bool,
-}
-
-impl RenderOnce for SshConnectionHeader {
-    fn render(self, _window: &mut Window, cx: &mut App) -> impl IntoElement {
-        let theme = cx.theme();
-
-        let mut header_color = theme.colors().text;
-        header_color.fade_out(0.96);
-
-        let (main_label, meta_label) = if let Some(nickname) = self.nickname {
-            (nickname, Some(format!("({})", self.connection_string)))
-        } else {
-            (self.connection_string, None)
-        };
-
-        let icon = if self.is_wsl {
-            IconName::Linux
-        } else if self.is_devcontainer {
-            IconName::Box
-        } else {
-            IconName::Server
-        };
-
-        h_flex()
-            .px(DynamicSpacing::Base12.rems(cx))
-            .pt(DynamicSpacing::Base08.rems(cx))
-            .pb(DynamicSpacing::Base04.rems(cx))
-            .rounded_t_sm()
-            .w_full()
-            .gap_1p5()
-            .child(Icon::new(icon).size(IconSize::Small))
-            .child(
-                h_flex()
-                    .gap_1()
-                    .overflow_x_hidden()
-                    .child(
-                        div()
-                            .max_w_96()
-                            .overflow_x_hidden()
-                            .text_ellipsis()
-                            .child(Headline::new(main_label).size(HeadlineSize::XSmall)),
-                    )
-                    .children(
-                        meta_label.map(|label| {
-                            Label::new(label).color(Color::Muted).size(LabelSize::Small)
-                        }),
-                    )
-                    .child(div().overflow_x_hidden().text_ellipsis().children(
-                        self.paths.into_iter().map(|path| {
-                            Label::new(path.to_string_lossy().into_owned())
-                                .size(LabelSize::Small)
-                                .color(Color::Muted)
-                        }),
-                    )),
-            )
-    }
-}
-
-impl Render for RemoteConnectionModal {
-    fn render(&mut self, window: &mut Window, cx: &mut Context<Self>) -> impl ui::IntoElement {
-        let nickname = self.prompt.read(cx).nickname.clone();
-        let connection_string = self.prompt.read(cx).connection_string.clone();
-        let is_wsl = self.prompt.read(cx).is_wsl;
-        let is_devcontainer = self.prompt.read(cx).is_devcontainer;
-
-        let theme = cx.theme().clone();
-        let body_color = theme.colors().editor_background;
-
-        v_flex()
-            .elevation_3(cx)
-            .w(rems(34.))
-            .border_1()
-            .border_color(theme.colors().border)
-            .key_context("SshConnectionModal")
-            .track_focus(&self.focus_handle(cx))
-            .on_action(cx.listener(Self::dismiss))
-            .on_action(cx.listener(Self::confirm))
-            .child(
-                SshConnectionHeader {
-                    paths: self.paths.clone(),
-                    connection_string,
-                    nickname,
-                    is_wsl,
-                    is_devcontainer,
-                }
-                .render(window, cx),
-            )
-            .child(
-                div()
-                    .w_full()
-                    .bg(body_color)
-                    .border_y_1()
-                    .border_color(theme.colors().border_variant)
-                    .child(self.prompt.clone()),
-            )
-            .child(
-                div().w_full().py_1().child(
-                    ListItem::new("li-devcontainer-go-back")
-                        .inset(true)
-                        .spacing(ui::ListItemSpacing::Sparse)
-                        .start_slot(Icon::new(IconName::Close).color(Color::Muted))
-                        .child(Label::new("Cancel"))
-                        .end_slot(
-                            KeyBinding::for_action_in(&menu::Cancel, &self.focus_handle(cx), cx)
-                                .size(rems_from_px(12.)),
-                        )
-                        .on_click(cx.listener(|this, _, window, cx| {
-                            this.dismiss(&menu::Cancel, window, cx);
-                        })),
-                ),
-            )
-    }
-}
-
-impl Focusable for RemoteConnectionModal {
-    fn focus_handle(&self, cx: &gpui::App) -> gpui::FocusHandle {
-        self.prompt.read(cx).editor.focus_handle(cx)
-    }
-}
-
-impl EventEmitter<DismissEvent> for RemoteConnectionModal {}
-
-impl ModalView for RemoteConnectionModal {
-    fn on_before_dismiss(
-        &mut self,
-        _window: &mut Window,
-        _: &mut Context<Self>,
-    ) -> workspace::DismissDecision {
-        workspace::DismissDecision::Dismiss(self.finished)
-    }
-
-    fn fade_out_background(&self) -> bool {
-        true
-    }
-}
-
-#[derive(Clone)]
-pub struct RemoteClientDelegate {
-    window: AnyWindowHandle,
-    ui: WeakEntity<RemoteConnectionPrompt>,
-    known_password: Option<EncryptedPassword>,
-}
-
-impl remote::RemoteClientDelegate for RemoteClientDelegate {
-    fn ask_password(
-        &self,
-        prompt: String,
-        tx: oneshot::Sender<EncryptedPassword>,
-        cx: &mut AsyncApp,
-    ) {
-        let mut known_password = self.known_password.clone();
-        if let Some(password) = known_password.take() {
-            tx.send(password).ok();
-        } else {
-            self.window
-                .update(cx, |_, window, cx| {
-                    self.ui.update(cx, |modal, cx| {
-                        modal.set_prompt(prompt, tx, window, cx);
-                    })
-                })
-                .ok();
-        }
-    }
-
-    fn set_status(&self, status: Option<&str>, cx: &mut AsyncApp) {
-        self.update_status(status, cx)
-    }
-
-    fn download_server_binary_locally(
-        &self,
-        platform: RemotePlatform,
-        release_channel: ReleaseChannel,
-        version: Option<Version>,
-        cx: &mut AsyncApp,
-    ) -> Task<anyhow::Result<PathBuf>> {
-        let this = self.clone();
-        cx.spawn(async move |cx| {
-            AutoUpdater::download_remote_server_release(
-                release_channel,
-                version.clone(),
-                platform.os.as_str(),
-                platform.arch.as_str(),
-                move |status, cx| this.set_status(Some(status), cx),
-                cx,
-            )
-            .await
-            .with_context(|| {
-                format!(
-                    "Downloading remote server binary (version: {}, os: {}, arch: {})",
-                    version
-                        .as_ref()
-                        .map(|v| format!("{}", v))
-                        .unwrap_or("unknown".to_string()),
-                    platform.os,
-                    platform.arch,
-                )
-            })
-        })
-    }
-
-    fn get_download_url(
-        &self,
-        platform: RemotePlatform,
-        release_channel: ReleaseChannel,
-        version: Option<Version>,
-        cx: &mut AsyncApp,
-    ) -> Task<Result<Option<String>>> {
-        cx.spawn(async move |cx| {
-            AutoUpdater::get_remote_server_release_url(
-                release_channel,
-                version,
-                platform.os.as_str(),
-                platform.arch.as_str(),
-                cx,
-            )
-            .await
-        })
-    }
-}
-
-impl RemoteClientDelegate {
-    fn update_status(&self, status: Option<&str>, cx: &mut AsyncApp) {
-        cx.update(|cx| {
-            self.ui
-                .update(cx, |modal, cx| {
-                    modal.set_status(status.map(|s| s.to_string()), cx);
-                })
-                .ok()
-        });
-    }
-}
-
-pub fn connect(
-    unique_identifier: ConnectionIdentifier,
-    connection_options: RemoteConnectionOptions,
-    ui: Entity<RemoteConnectionPrompt>,
-    window: &mut Window,
-    cx: &mut App,
-) -> Task<Result<Option<Entity<RemoteClient>>>> {
-    let window = window.window_handle();
-    let known_password = match &connection_options {
-        RemoteConnectionOptions::Ssh(ssh_connection_options) => ssh_connection_options
-            .password
-            .as_deref()
-            .and_then(|pw| pw.try_into().ok()),
-        _ => None,
-    };
-    let (tx, mut rx) = oneshot::channel();
-    ui.update(cx, |ui, _cx| ui.set_cancellation_tx(tx));
-
-    let delegate = Arc::new(RemoteClientDelegate {
-        window,
-        ui: ui.downgrade(),
-        known_password,
-    });
-
-    cx.spawn(async move |cx| {
-        let connection = remote::connect(connection_options, delegate.clone(), cx);
-        let connection = select! {
-            _ = rx => return Ok(None),
-            result = connection.fuse() => result,
-        }?;
-
-        cx.update(|cx| remote::RemoteClient::new(unique_identifier, connection, rx, delegate, cx))
-            .await
-    })
-}
-
 pub async fn open_remote_project(
     connection_options: RemoteConnectionOptions,
     paths: Vec<PathBuf>,
@@ -641,7 +135,6 @@ pub async fn open_remote_project(
     } else {
         let workspace_position = cx
             .update(|cx| {
-                // todo: These paths are wrong they may have column and line information
                 workspace::remote_workspace_position_from_db(connection_options.clone(), &paths, cx)
             })
             .await
@@ -695,12 +188,10 @@ pub async fn open_remote_project(
                     ui.set_cancellation_tx(cancel_tx);
                 });
 
-                Some(Arc::new(RemoteClientDelegate {
-                    window: window.window_handle(),
-                    ui: ui.downgrade(),
-                    known_password: if let RemoteConnectionOptions::Ssh(options) =
-                        &connection_options
-                    {
+                Some(Arc::new(RemoteClientDelegate::new(
+                    window.window_handle(),
+                    ui.downgrade(),
+                    if let RemoteConnectionOptions::Ssh(options) = &connection_options {
                         options
                             .password
                             .as_deref()
@@ -708,7 +199,7 @@ pub async fn open_remote_project(
                     } else {
                         None
                     },
-                }))
+                )))
             }
         })?;
 
@@ -884,7 +375,6 @@ pub async fn open_remote_project(
             }
         })
         .ok();
-    // Already showed the error to the user
     Ok(())
 }
 
@@ -933,9 +423,9 @@ async fn path_exists(connection: &Arc<dyn RemoteConnection>, path: &Path) -> boo
 #[cfg(test)]
 mod tests {
     use super::*;
-    use extension::ExtensionHostProxy;
+    use extension_host::ExtensionHostProxy;
     use fs::FakeFs;
-    use gpui::TestAppContext;
+    use gpui::{AppContext, TestAppContext};
     use http_client::BlockedHttpClient;
     use node_runtime::NodeRuntime;
     use remote::RemoteClient;
@@ -948,83 +438,94 @@ mod tests {
         cx: &mut TestAppContext,
         server_cx: &mut TestAppContext,
     ) {
-        let app_state = init_test(cx);
-        let executor = cx.executor();
-
-        cx.update(|cx| {
-            release_channel::init(semver::Version::new(0, 0, 0), cx);
-        });
-        server_cx.update(|cx| {
-            release_channel::init(semver::Version::new(0, 0, 0), cx);
-        });
-
-        let (opts, server_session, connect_guard) = RemoteClient::fake_server(cx, server_cx);
-
-        let remote_fs = FakeFs::new(server_cx.executor());
-        remote_fs
-            .insert_tree(
-                path!("/project"),
-                json!({
+        let fs = FakeFs::new(server_cx.executor());
+        fs.insert_tree(
+            path!("/code"),
+            json!({
+                "project1": {
+                    "README.md": "# Project 1",
                     "src": {
-                        "main.rs": "fn main() {}",
-                    },
-                    "README.md": "# Test Project",
-                }),
-            )
-            .await;
+                        "lib.rs": "fn hello() {}",
+                    }
+                },
+                "project2": {
+                    "README.md": "# Project 2",
+                }
+            }),
+        )
+        .await;
 
-        server_cx.update(HeadlessProject::init);
-        let http_client = Arc::new(BlockedHttpClient);
-        let node_runtime = NodeRuntime::unavailable();
-        let languages = Arc::new(language::LanguageRegistry::new(server_cx.executor()));
-        let proxy = Arc::new(ExtensionHostProxy::new());
+        let server_ssh = remote::SshRemoteProxy::fake_server(Default::default(), server_cx).await;
+        let client_ssh = remote::SshRemoteProxy::fake_client(server_ssh.clone(), cx).await;
 
-        let _headless = server_cx.new(|cx| {
+        let headless_project = server_cx.new(|cx| {
             HeadlessProject::new(
                 HeadlessAppState {
-                    session: server_session,
-                    fs: remote_fs.clone(),
-                    http_client,
-                    node_runtime,
-                    languages,
-                    extension_host_proxy: proxy,
+                    fs: fs.clone(),
+                    http_client: Arc::new(BlockedHttpClient),
+                    node_runtime: NodeRuntime::unavailable(),
+                    languages: Arc::new(language::LanguageRegistry::new(
+                        server_cx.executor(),
+                        server_cx.executor(),
+                    )),
+                    extension_host_proxy: ExtensionHostProxy::uninitialized(),
                 },
-                false,
+                server_ssh.incoming_rx.lock().take().unwrap(),
+                cx,
+            )
+        });
+        server_cx.run_until_parked();
+
+        let remote_client = cx.update(|cx| {
+            let (cancel_tx, cancel_rx) = oneshot::channel();
+            drop(cancel_tx);
+            RemoteClient::new(
+                remote::ConnectionIdentifier::setup(),
+                Arc::new(client_ssh),
+                cancel_rx,
+                Arc::new(remote::test::NoOpRemoteClientDelegate),
                 cx,
             )
         });
 
-        drop(connect_guard);
+        cx.run_until_parked();
+        server_cx.run_until_parked();
 
-        let paths = vec![PathBuf::from(path!("/project"))];
-        let open_options = workspace::OpenOptions::default();
+        let remote_client = remote_client.await.unwrap().unwrap();
 
-        let mut async_cx = cx.to_async();
-        let result = open_remote_project(opts, paths, app_state, open_options, &mut async_cx).await;
+        let project = cx.update(|cx| {
+            project::Project::remote(
+                remote_client,
+                Arc::new(client::Client::new(
+                    Arc::new(BlockedHttpClient {}),
+                    Arc::new(http_client::Url::parse("https://localhost").unwrap()),
+                    cx,
+                )),
+                NodeRuntime::unavailable(),
+                Arc::new(client::UserStore::new(cx)),
+                Arc::new(language::LanguageRegistry::new(
+                    cx.background_executor().clone(),
+                    cx.background_executor().clone(),
+                )),
+                Arc::new(fs::RealFs::default()),
+                true,
+                cx,
+            )
+        });
 
-        executor.run_until_parked();
-
-        assert!(result.is_ok(), "open_remote_project should succeed");
-
-        let windows = cx.update(|cx| cx.windows().len());
-        assert_eq!(windows, 1, "Should have opened a window");
-
-        let workspace_handle = cx.update(|cx| cx.windows()[0].downcast::<Workspace>().unwrap());
-
-        workspace_handle
-            .update(cx, |workspace, _, cx| {
-                let project = workspace.project().read(cx);
-                assert!(project.is_remote(), "Project should be a remote project");
+        project
+            .update(cx, |project, cx| {
+                project.find_or_create_worktree(path!("/code/project1"), true, cx)
             })
+            .await
             .unwrap();
+
+        cx.run_until_parked();
+        server_cx.run_until_parked();
+
+        drop(project);
+        drop(headless_project);
     }
 
-    fn init_test(cx: &mut TestAppContext) -> Arc<AppState> {
-        cx.update(|cx| {
-            let state = AppState::test(cx);
-            crate::init(cx);
-            editor::init(cx);
-            state
-        })
-    }
+    fn init_test(_: &mut TestAppContext) {}
 }
